feature: add totalPoint on each collect

+++ src/main/java/com/example/repick/domain/clothingSales/dto/GetSellingClothingSales.java
@@ -10,6 +10,6 @@ public record GetSellingClothingSales(
         @Schema(description = "판매 중인 의류 수량", example = "43") Integer sellingQuantity,
         @Schema(description = "구매 확정 대기 수량", example = "7") Integer pendingQuantity,
         @Schema(description = "구매 확정 완료 수량", example = "36") Integer soldQuantity,
-        @Schema(description = "총 포인트", example = "2300") Integer totalPoint
+        @Schema(description = "총 포인트", example = "2300") Long totalPoint
 ) {
 }
